Handle multi-recipient thread titles
The ThreadTitle component is refactored out into two components,
Recipients and ThreadSubject.

The Recipients component handles two types of recipients:
A string (which is the typical case) and an array of recipients
(for example, in a private message).

When recieving an array of recipients, if there is only one recipient
and the recipient is the current user, the string displayed is "you".

If there is only one recipient and they are not the current user, the string
displayed is the name of the user.

If there are more than one recipients, e.g., [A, B, C], the string
displayed is "A, B and C"

The ThreadSubject component returns a dot followed by the subject
if there is a subject. Otherwise, it returns nothing (no dot).

The dot style object is now moved to styles/threadSubject.

This fixes #17